Adding guide system (part 2)
- Adding guide grid controls
- Adding guide line controls
- Removing textdraw margin

The textdraw margin is not scaled, so it was removed from the guide system.

If you need to create a box with a scaled margin, it is recommended to use font 4 and the sprite ld_bum:blkdot.

+++ assets/js/ui/dialogui/guidelinedialogui.js
@@ -23,7 +23,7 @@ function GuideLineDialogUI(parent, title, x, y, size, padding, style, clickAccep
 	this.styleUI.appendOption("Vertical");
 	this.styleUI.element.selectedIndex = style;
 	this.contentUI.appendStaticLine();
-	this.buttonAcceptUI = new ButtonUI(this.buttonsUI, {innerText: "Accept", click: () => { clickAccept(parseFloat(this.xUI.element.value), parseFloat(this.yUI.element.value), parseFloat(this.sizeUI.element.value), parseInt(this.paddingUI.element.value), this.styleUI.element.selectedIndex); }});
+	this.buttonAcceptUI = new ButtonUI(this.buttonsUI, {innerText: "Accept", click: () => { clickAccept(parseFloat(this.xUI.element.value), parseFloat(this.yUI.element.value), parseFloat(this.sizeUI.element.value), parseFloat(this.paddingUI.element.value), this.styleUI.element.selectedIndex); }});
 	this.buttonCancelUI = new ButtonUI(this.buttonsUI, {innerText: "Cancel", click: () => { clickCancel(); }});
 }
